Reimplement command runner using tokio.

+++ src/run.rs
@@ -1,14 +1,20 @@
+#![deny(unused_must_use)]
+
 use errors::*;
+use futures::{Future, Stream};
 use slog::Logger;
 use slog_scope;
-use std::io::{BufRead, BufReader, Read};
+use std::cell::Cell;
+use std::io::{self, BufReader};
 use std::ops::Deref;
 use std::path::Path;
 use std::process::{Command, ExitStatus, Stdio};
-use std::sync::mpsc::{self, RecvTimeoutError};
-use std::sync::mpsc::{Receiver, Sender, channel};
-use std::thread;
+use std::rc::Rc;
 use std::time::Duration;
+use tokio_core::reactor::Core;
+use tokio_io::io::lines;
+use tokio_process::CommandExt;
+use tokio_timer;
 
 pub fn run(name: &str, args: &[&str], env: &[(&str, &str)]) -> Result<()> {
     run_full(None, name, args, env)?;
@@ -93,22 +99,44 @@ const MAX_TIMEOUT_SECS: u64 = 60 * 10 * 2;
 const HEARTBEAT_TIMEOUT_SECS: u64 = 60 * 2;
 
 fn log_command_(mut cmd: Command, capture: bool) -> Result<ProcessOutput> {
-    let mut child = cmd.stdout(Stdio::piped()).stderr(Stdio::piped()).spawn()?;
+    let mut core = Core::new().unwrap();
+    let timer = tokio_timer::wheel()
+        .max_timeout(Duration::from_secs(MAX_TIMEOUT_SECS * 2))
+        .build();
+    let mut child = cmd.stdout(Stdio::piped())
+        .stderr(Stdio::piped())
+        .spawn_async(&core.handle())?;
 
-    let stdout = child.stdout.take().expect("");
-    let stderr = child.stderr.take().expect("");
+    let stdout = child.stdout().take().expect("");
+    let stderr = child.stderr().take().expect("");
 
     // Needed for killing after timeout
     let child_id = child.id();
 
-    // Child's stdio needs to produce output to avoid being killed
-    let (heartbeat_tx, heartbeat_rx) = mpsc::channel();
+    let heartbeat_timeout = Duration::from_secs(MAX_TIMEOUT_SECS);
 
-    let rx_out = sink(Box::new(stdout),
-                      log_child_stdout,
-                      capture,
-                      heartbeat_tx.clone());
-    let rx_err = sink(Box::new(stderr), log_child_stderr, capture, heartbeat_tx);
+    let heartbeat_timed_out = Rc::new(Cell::new(false));
+    fn kill_on_heartbeat_timeout(child_id: u32, cond: &Rc<Cell<bool>>, e: io::Error) -> io::Error {
+        if e.kind() == io::ErrorKind::TimedOut {
+            kill_process(child_id);
+            cond.set(true);
+        }
+        e
+    };
+    let rx_out = timer
+        .timeout_stream(lines(BufReader::new(stdout)), heartbeat_timeout)
+        .map_err({
+                     let cond = heartbeat_timed_out.clone();
+                     move |e| kill_on_heartbeat_timeout(child_id, &cond, e)
+                 });
+    let rx_err = timer
+        .timeout_stream(lines(BufReader::new(stderr)), heartbeat_timeout)
+        .map_err({
+                     let cond = heartbeat_timed_out.clone();
+                     move |e| kill_on_heartbeat_timeout(child_id, &cond, e)
+                 });
+    let rx_out = sink(rx_out, log_child_stdout, capture);
+    let rx_err = sink(rx_err, log_child_stderr, capture);
 
     #[cfg(unix)]
     fn kill_process(id: u32) {
@@ -129,64 +157,26 @@ fn log_command_(mut cmd: Command, capture: bool) -> Result<ProcessOutput> {
         };
     }
 
-    // Have another thread kill the subprocess after the maximum timeout
-    let (timeout_tx, timeout_rx) = mpsc::channel();
-    let (timeout_cancel_tx, timeout_cancel_rx) = mpsc::channel();
-    thread::spawn(move || {
-        let timeout = Duration::from_secs(MAX_TIMEOUT_SECS);
-        match timeout_cancel_rx.recv_timeout(timeout) {
-            Ok(_) => {
-                // Child process exited
-                return;
-            }
-            Err(RecvTimeoutError::Timeout) => {
-                kill_process(child_id);
-                timeout_tx.send(());
-            }
-            _ => panic!(),
-        }
-    });
+    let timed_out = Cell::new(false);
+    let child = timer
+        .timeout(child, Duration::from_secs(MAX_TIMEOUT_SECS))
+        .map_err(|e| {
+                     if e.kind() == io::ErrorKind::TimedOut {
+                         kill_process(child_id);
+                         timed_out.set(true);
+                     }
+                     e
+                 });
 
-    // Have another thread listening for heartbeats on stdout/stderr
-    let (heartbeat_timeout_tx, heartbeat_timeout_rx) = mpsc::channel();
-    let (heartbeat_cancel_tx, heartbeat_cancel_rx) = mpsc::channel();
-    thread::spawn(move || {
-        loop {
-            let timeout = Duration::from_secs(HEARTBEAT_TIMEOUT_SECS);
-            match heartbeat_cancel_rx.recv_timeout(timeout) {
-                Ok(_) => {
-                    // Child process exited
-                    return;
-                }
-                Err(RecvTimeoutError::Timeout) => {
-                    let heartbeats = heartbeat_rx.try_iter().count();
-                    if heartbeats > 0 {
-                        continue;
-                    } else {
-                        // No heartbeats before timeout
-                        kill_process(child_id);
-                        heartbeat_timeout_tx.send(());
-                    }
-                }
-                _ => panic!(),
-            }
-        }
-    });
 
-    let status = child.wait();
-    timeout_cancel_tx.send(());
-    heartbeat_cancel_tx.send(());
-    let timed_out = timeout_rx.try_recv().is_ok();
-    let heartbeat_timed_out = heartbeat_timeout_rx.try_recv().is_ok();
-    let status = status?;
-    let stdout = rx_out.recv().expect("");
-    let stderr = rx_err.recv().expect("");
+    // TODO: Handle errors from tokio_timer better, in particular TimerError::TooLong
+    let (status, stdout, stderr) = core.run(Future::join3(child, rx_out, rx_err))?;
 
-    if heartbeat_timed_out {
+    if heartbeat_timed_out.get() {
         info!("process killed after not generating output for {} s",
               HEARTBEAT_TIMEOUT_SECS);
         bail!(ErrorKind::Timeout);
-    } else if timed_out {
+    } else if timed_out.get() {
         info!("process killed after max time of {} s", MAX_TIMEOUT_SECS);
         bail!(ErrorKind::Timeout);
     }
@@ -204,33 +194,23 @@ fn log_child_stdout(logger: &Logger, line: &str) {
 
 fn log_child_stderr(logger: &Logger, line: &str) {
     slog_info!(logger, "kablam! {}", line);
+    println!("{}", line);
 }
 
-fn sink(reader: Box<Read + Send>,
-        log: fn(&Logger, &str),
-        capture: bool,
-        heartbeat_tx: Sender<()>)
-        -> Receiver<Vec<String>> {
-    let (tx, rx) = channel();
+fn sink<S>(reader: S,
+           log: fn(&Logger, &str),
+           capture: bool)
+           -> Box<Future<Item = Vec<String>, Error = io::Error>>
+    where S: Stream<Item = String, Error = io::Error> + 'static
+{
     let logger = slog_scope::logger();
-    thread::spawn(move || {
-        let mut buf = Vec::new();
-        let reader = BufReader::new(reader);
-        for line_bytes in reader.split(b'\n') {
-            if let Ok(line_bytes) = line_bytes {
-                let line = String::from_utf8_lossy(&line_bytes);
-                log(&logger, line.deref());
-                heartbeat_tx.send(());
-                if capture {
-                    buf.push(line.to_string());
-                }
-            } else {
-                log(&logger, "READING FROM CHILD PROCESS FAILED!");
-            }
-        }
-
-        tx.send(buf).expect("");
-    });
-
-    rx
+    let reader = reader.map(move |line| {
+                                log(&logger, line.deref());
+                                line.to_string()
+                            });
+    if capture {
+        Box::new(reader.collect())
+    } else {
+        Box::new(reader.for_each(|_| Ok(())).map(|_| Vec::new()))
+    }
 }
